fix(frontend_auth): add steppers to login authentication
