+++ frontend/src/app/[locale]/(auth)/signin/page.tsx
@@ -1,5 +1,5 @@
 import { auth } from "@/auth";
-import SignInForm from "@/components/Auth/SignInForm";
+import SignInForm from "@/components/Auth/SignIn/SignInForm";
 import { redirect } from "next/navigation";
 
 const SigninPage = async () => {
